device_scripts/requirements.txt
awsiotsdk
adafruit-circuitpython-dht
getmac
gpiozero
spidev
